requirements.txt
# Core dependencies
pipecat-ai[cartesia,anthropic,silero,websocket]>=0.0.90
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0

# Configuration and env
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# MCP (using mcp-use SDK for hackathon compatibility)
mcp-use>=1.6.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
aiofiles>=23.0.0
tenacity>=8.2.0
